Update mymath.go
float64

+++ mymath.go
@@ -2,7 +2,7 @@ package mymath
 
 import "math"
 
-func Sqrt(x float64) int {
+func Sqrt(x float64) float64 {
 	return math.Sqrt(x)
 }
